Add files via upload
Problem Statement:

A Chinese automobile company Geely Auto aspires to enter the US market by setting up their manufacturing unit there and producing cars locally to give competition to their US and European counterparts. They have contracted an automobile consulting company since they want to understand the factors on which pricing of a car depends. Essentially the company wants to know:
•	Which variables are highly significant in estimating the price of a car
•	How well those important variables describe the price of a car
Based on various market surveys, the consulting firm has gathered a large dataset of different types of cars across the world and their attributes.

Objective:

You are required to model the price of cars with the available independent variables. It will be used by the management to understand how exactly the prices vary with the independent variables. Further, the model will be a good way for the management to understand the pricing dynamics of a new market.