Add files via upload
Problem Statement:

A large company named XYZ, employs, at any given point of time, around 4000 employees. However, every year, around 15% of its employees leave the company and need to be replaced with the talent pool available in the job market. Hence, the management has contracted an analytics firm to understand what factors they should focus on, in order to curb attrition. Also, they want to know which of these variables is most important and needs to be addressed right away.

Objective:

You are required to model the probability of attrition using a logistic regression model. The results thus obtained will be used by the management to understand what changes they should make to their workplace, in order to get most of their employees to stay.