Add files via upload
Problem statement:

A classic problem in the field of pattern recognition is that of handwritten digit recognition. Suppose that you have an image of a digit submitted by a user via a scanner, a tablet, or other digital devices. The goal is to develop a model that can correctly identify the digit (between 0-9) written in an image. 
 
Objective:

You are required to develop a model using Support Vector Machine which should correctly classify the handwritten digits based on the pixel values given as features.